StopForumSpam API
This calls the StopForumSpam API at stopforumspam.org. As per day 20000
calls are allowed from a shared host it's nice that the calls can be
restricted to 1 or 2 calls per actual subscriber, first on IP, second on
email, as spammer IP-addresses are less volatile. The combination
filtered all 100 subscribers correct in a one week run.